# docs/api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Employee Management System API
  description: |
    RESTful API for managing employees, users, leave requests, and authentication.
    
    Features:
    - Employee CRUD operations with CSV import/export
    - User management with role-based access control
    - Leave request workflow with approval/rejection
    - Email verification and password reset
    - Rate limiting and caching
  version: 1.0.0
  contact:
    name: EMS Development Team
  license:
    name: MIT License
servers:
  - url: https://employee-management-system-production-2a1a.up.railway.app/api/v1
    description: Production server (Railway)
  - url: http://localhost:8080/api/v1
    description: Local development server
security:
  - cookieAuth: []
...
paths:
...
  /leave/balance/{employeeId}:
    get:
      tags:
        - Leave Management
      summary: Get leave balance
      operationId: getLeaveBalance
      security:
        - cookieAuth: []
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Leave balance retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveBalance'
  /leave/request:
    post:
      tags:
        - Leave Management
      summary: Create leave request
      operationId: createLeaveRequest
      security:
        - cookieAuth: []
      parameters:
        - name: employeeId
          in: query
          required: true
          schema:
            type: integer
        - name: leaveType
          in: query
          required: true
          schema:
            type: string
            enum: [ANNUAL, SICK, EMERGENCY]
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: reason
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Leave request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveRequest'
  /leave/requests/{requestId}/approve:
    post:
      tags:
        - Leave Management
      summary: Approve leave request
      operationId: approveLeaveRequest
      security:
        - cookieAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Leave request approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveRequest'
  /leave/requests/{requestId}/reject:
    post:
      tags:
        - Leave Management
      summary: Reject leave request
      operationId: rejectLeaveRequest
      security:
        - cookieAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
        - name: rejectionReason
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Leave request rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveRequest'

  # Password Reset
  /password-reset/request:
    post:
      tags:
        - Password Reset
      summary: Request password reset
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Password reset email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /password-reset/validate:
    post:
      tags:
        - Password Reset
      summary: Validate reset code
      operationId: validateResetToken
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Code is valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /password-reset/confirm:
    post:
      tags:
        - Password Reset
      summary: Confirm password reset
      operationId: confirmPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                newPassword:
                  type: string
                  minLength: 6
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  # Email Verification
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Session-based authentication
  schemas:
...
    AuthResponse:
      type: object
      properties:
        message:
          type: string
          example: "Login successful"
        success:
          type: boolean
          example: true
...
    LeaveBalance:
      type: object
      properties:
        id:
          type: integer
        annualLeave:
          type: integer
          example: 20
        emergencyLeave:
          type: integer
          example: 5
        sickLeave:
          type: integer
          example: 10
    LeaveRequest:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        leaveType:
          type: string
          enum: [ANNUAL, SICK, EMERGENCY]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        daysCount:
          type: integer
        reason:
          type: string
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        approvedBy:
          type: string
        approvedAt:
          type: string
          format: date-time
        rejectionReason:
          type: string
...